chore: remove second page of user operator form
