+++ client/app/components/button/RequestAccessButton.tsx
@@ -41,14 +41,7 @@ export default function RequestAccessButton({
       setErrorList([{ message: response.error }]);
       return;
     }
-
-    // if user is requesting access as an admin, redirect to user operator form
-    if (isAdminRequest)
-      push(
-        `/dashboard/select-operator/user-operator/request-access/2?user-operator-id=${response.user_operator_id}`,
-      );
-    else
-      push(`/dashboard/select-operator/received/request-access/${operatorId}`);
+    push(`/dashboard/select-operator/received/request-access/${operatorId}`);
   };
 
   return (
