+++ client/app/components/form/UserOperatorForm.tsx
@@ -1,0 +1,136 @@
+"use client";
+
+import { RJSFSchema } from "@rjsf/utils";
+import { useState } from "react";
+import {
+  useParams,
+  usePathname,
+  useRouter,
+  useSearchParams,
+} from "next/navigation";
+import { userOperatorUiSchema } from "@/app/utils/jsonSchema/userOperator";
+import { actionHandler } from "@/app/utils/actions";
+import { useSession } from "next-auth/react";
+import UserOperatorReview from "@/app/components/routes/access-requests/form/UserOperatorReview";
+import MultiStepFormBase from "@/app/components/form/MultiStepFormBase";
+import { UserOperatorFormData } from "@/app/components/form/formDataTypes";
+import Note from "../datagrid/Note";
+import { Status, UserOperatorStatus } from "@/app/utils/enums";
+
+interface UserOperatorFormProps {
+  readonly schema: RJSFSchema;
+  readonly formData: Partial<UserOperatorFormData>;
+}
+export default function UserOperatorForm({
+  schema,
+  formData,
+}: UserOperatorFormProps) {
+  const { data: session } = useSession();
+  const { push } = useRouter();
+  const pathname = usePathname();
+  const params = useParams();
+  const searchParams = useSearchParams();
+  const [error, setError] = useState(undefined);
+  const formSection =
+    parseInt(params?.formSection as string) ||
+    parseInt(searchParams.get("form-section") as string);
+  const isCreate = pathname.includes("create");
+  const isRequestingAccess = pathname.includes("request");
+
+  const userOperatorId = params?.id as string;
+  const submitHandler = async (data: { formData?: UserOperatorFormData }) => {
+    const newFormData = {
+      ...data.formData,
+    } as UserOperatorFormData;
+
+    const response = await actionHandler(
+      `registration/user-operator/operator${
+        !isCreate ? `/${userOperatorId}` : ""
+      }`,
+      isCreate ? "POST" : "PUT",
+      `/dashboard/select-operator/user-operator/${
+        isCreate ? "create" : userOperatorId
+      }/${formSection}`,
+      {
+        body: JSON.stringify(newFormData),
+      },
+    );
+    if (response.error) {
+      setError(response.error);
+      // return error so MultiStepFormBase can re-enable the submit button
+      // and user can attempt to submit again
+      return { error: response.error };
+    }
+    if (isCreate) {
+      return push(
+        `/dashboard/select-operator/received/add-operator/${response.operator_id}`,
+      );
+    }
+
+    if (isRequestingAccess) {
+      return push(
+        `/dashboard/select-operator/received/request-access/${response.operator_id}`,
+      );
+    }
+    return push(`/dashboard`);
+  };
+
+  const isCasInternal =
+    session?.user.app_role?.includes("cas") &&
+    !session?.user.app_role?.includes("pending");
+
+  const isFormStatusDisabled =
+    formData?.status === Status.PENDING || formData?.status === Status.APPROVED;
+
+  const isNewOperatorMessage = formData.is_new && formSection === 1;
+  const operatorRoute = isCasInternal ? "operators" : "select-operator";
+
+  return (
+    <>
+      {!isCasInternal && formData.status === UserOperatorStatus.DRAFT && (
+        <Note
+          classNames={"mb-4 mt-6"}
+          showNotePrefix={false}
+          showAlertIcon={true}
+          message="Please fill out missing information or update incorrect information about your operator to continue with your application.
+Some fields cannot be edited. If you need to change those fields, please contact us via email at GHGRegulator@gov.bc.ca."
+        />
+      )}
+
+      {isCasInternal && (
+        <>
+          {isNewOperatorMessage && (
+            <Note message="This is a new operator. You must approve this operator before approving its admin." />
+          )}
+          <UserOperatorReview
+            userOperator={formData as UserOperatorFormData}
+            userOperatorId={Number(userOperatorId)}
+            isOperatorNew={formData?.is_new}
+            operatorId={formData?.operator_id}
+            // We don't want to show the request changes button for Prime Admin approval
+            showRequestChanges={false}
+          />
+        </>
+      )}
+
+      <MultiStepFormBase
+        cancelUrl={
+          isCasInternal ? "/dashboard/operators" : "/dashboard/select-operator"
+        }
+        allowEdit={isFormStatusDisabled && !isCasInternal}
+        allowBackNavigation
+        baseUrl={`/dashboard/${operatorRoute}/user-operator/${
+          isCreate ? "create" : userOperatorId
+        }`}
+        schema={schema}
+        disabled={isCasInternal || isFormStatusDisabled}
+        error={error}
+        setErrorReset={setError}
+        formData={formData}
+        onSubmit={submitHandler}
+        submitButtonText={isCreate ? "Submit" : "Save and Return to Dashboard"}
+        uiSchema={userOperatorUiSchema}
+      />
+    </>
+  );
+}
